# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/clients/{clientId}/shifts": {
      "get": {
        "operationId": "ShiftsController_list",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar horários de acesso do cliente",
        "tags": [
          "shifts"
        ]
      },
      "post": {
        "operationId": "ShiftsController_create",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar horário",
        "tags": [
          "shifts"
        ]
      }
    },
    "/clients/{clientId}/shifts/{shiftId}": {
      "get": {
        "operationId": "ShiftsController_getOne",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe do horário",
        "tags": [
          "shifts"
        ]
      },
      "patch": {
        "operationId": "ShiftsController_update",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar horário",
        "tags": [
          "shifts"
        ]
      },
      "delete": {
        "operationId": "ShiftsController_remove",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover horário",
        "tags": [
          "shifts"
        ]
      }
    },
...
    "/clients/{clientId}/responsibles": {
      "get": {
        "operationId": "ResponsiblesController_list",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar responsáveis da escola paginados (?page, ?pageSize, ?search)",
        "tags": [
          "responsibles"
        ]
      },
      "post": {
        "operationId": "ResponsiblesController_create",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cadastrar responsável com login (usuário + senha)",
        "tags": [
          "responsibles"
        ]
      }
    },
    "/clients/{clientId}/responsibles/face/global-sync/progress": {
      "get": {
        "operationId": "ResponsiblesController_globalSyncProgress",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "SSE — progresso da sincronização global de responsáveis (token na query)",
        "tags": [
          "responsibles"
        ]
      }
    },
    "/clients/{clientId}/responsibles/{responsibleId}/students": {
      "get": {
        "operationId": "ResponsiblesController_listStudents",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar alunos vinculados ao responsável",
        "tags": [
          "responsibles"
        ]
      },
      "post": {
        "operationId": "ResponsiblesController_linkStudent",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vincular aluno ao responsável",
        "tags": [
          "responsibles"
        ]
      }
    },
    "/clients/{clientId}/responsibles/{responsibleId}/students/{studentId}": {
      "delete": {
        "operationId": "ResponsiblesController_unlinkStudent",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover vínculo aluno–responsável",
        "tags": [
          "responsibles"
        ]
      },
      "patch": {
        "operationId": "ResponsiblesController_updateLink",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar vínculo (parentesco / autorização de retirada)",
        "tags": [
          "responsibles"
        ]
      }
    },
    "/clients/{clientId}/responsibles/{responsibleId}/face/sync": {
      "post": {
        "operationId": "ResponsiblesController_syncFace",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sincronizar face do responsável com os leitores Intelbras",
        "tags": [
          "responsibles"
        ]
      }
    },
    "/clients/{clientId}/responsibles/{responsibleId}": {
      "get": {
        "operationId": "ResponsiblesController_getOne",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe do responsável",
        "tags": [
          "responsibles"
        ]
      },
      "patch": {
        "operationId": "ResponsiblesController_update",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar responsável (opcional: nova senha)",
        "tags": [
          "responsibles"
        ]
      },
      "delete": {
        "operationId": "ResponsiblesController_deleteResponsible",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir responsável (remove face, placas, veículos e vínculos)",
        "tags": [
          "responsibles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Face2go API",
    "description": "Backend NestJS da plataforma Face2go",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}